Add guide for local TMT debugging
It's harder to debug/develop TMT locally because if you start VM then
you need to clone the repository for testing into the VM. That means
that the code for testing needs to be pushed including test scripts.

For that reason it's easier to go somewhere in the middle ground. Use
PR for your changes to the scripts and code which will allow you to use
local TMT plan changes.

However, good thing is that the TMT plan shouldn't be required to change
so often.